Reworked Aria to support multiple test folders.
